[Target_Core_Mod/ALUA]: Make EVPD 0x83 identifers return GOOD status during overflow
This patch makes transport_generic_emulate_inquiry() properly handle the case
where a received allocation length is smaller than the total outgoing EVPD 0x83
DataIN payload containing the full set of emulated identifiers.  This is what
scsi_dh_alua does; use an allocation length of 36 and then retry with a larger
allocation length based on what is returned in the EVPD payload during the initial
overflow case.

Instead of failing during the overflow case as before, this patch will calculate
the total size required to transmit the complete set of emulated EVPD 0x83
identifiers, and encode this value into the response payload with GOOD SAM status.
At this point the Initiator Port is expected to notice that the encoded EVPD length
is greater than the allocation length of the original CDB, and retry again with new
allocation length in order to retreive the complete set of EVPD 0x83 identifiers.

Signed-off-by: Nicholas A. Bellinger <nab@linux-iscsi.org>