[Target_Core_Mod]: Change all 'evpd' code references to 'vpd'
This patch converts all usage of the term 'evpd' and 'EVPD' to
'vpd' and 'VPD' respectively in target_core_mod code.  This change was made
because spc4 documents explictly list the device identification page (0x83)
as 'vital product data', and not 'extended vital product data'.

This patch includes changes to evpd_* prefixed ConfigFS attributes in
/sys/kernel/config/target/core/$HBA/$DEV/wwn/, as well as t10_evpd_t
structure and a number of other function names, defines, and printk() output.

Signed-off-by: Nicholas A. Bellinger <nab@linux-iscsi.org>