libfc: fix sequence-initiative WARN in fc_seq_start_next
The target code was getting a warning that sequence initiative
wasn't held when starting the response sequence after sending
the data sequence.

The bug was that sequence initiative was cleared due to the
END_SEQ flag being on.  The intent may have been to check LAST_SEQ.
Change just to check SEQ_INIT.

Signed-off-by: Joe Eykholt <jeykholt@cisco.com>
Committed-by: Nicholas A. Bellinger <nab@linux-iscsi.org>