tcm: Move remaining structure member related defs into related structures
This patch moves a handful of the remaining #defines from the top of
target_core_base.h inside of the structs and near the members that the
defines are actually related to.  Note that the #defines that have been
left in place are only used externally, and not directly related to a
particular structure member defined in target_core_base.h

Thanks to Konrad Rzeszutek Wilk for making this suggestion!

Signed-off-by: Nicholas A. Bellinger <nab@linux-iscsi.org>