target: remove SHUTDOWN_SIGS
In the these days with the kthread API we should never send signals to
kernel threads, but let start/stop be handled by the kthread helpers.

Note that the set_user_nice for all threads in the target code does not
seem particularly nice to the rest of the system.  Please re-evaluate if
the nice value is nessecary.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Acked-by: Nicholas A. Bellinger <nab@linux-iscsi.org>