target: take advantage of REQ_FUA in the iblock backend
Since 2.6.37-rc barriers have been replaced with the REQ_FLUSH and
REQ_FUA flags.  Both flags work on all devices without a need to check
for barrier functionality.

Use this fact to simplify the iblock code.  Use the REQ_FUA flag when
we report a write cache or if the target set the FUA bit, and remove
all the barrier detection code.

Btw, I think iblock would benefit from sharing a common option with
the file backend for enabling or disabling the option to provide
a voltile write cache to the initiator or not.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Acked-by: Nicholas A. Bellinger <nab@linux-iscsi.org>