target: simplify memory allocation
Whether we allocate a single buffer, an S/G list or nothing really only
depends on the command, and not the backend.  Unify all the allocation
code in a central place, keying of the cmd flags telling if it's an
S/G or linear buffer.  These were matching the function pointers in
all cases except for the WRITE SAME with unmap bit emulation, but that
one doesn't actually look at the payload anyway.

Handle the memory mapped ramdisk special inside the common code by
checking for the presence of a do_se_mem_map method.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Acked-by: Nicholas A. Bellinger <nab@linux-iscsi.org>