target: simplify subsystem registration and lookup
Move the subsystem registration and lookup into target_core_hba.c, and
make the list of subsystems local to the file.  Get rid of the sub_api_hba_cnt
field and instead to proper reference counting of the module inside the
subsystem_mutex to make it non-racy vs module removal.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Nicholas A. Bellinger <nab@linux-iscsi.org>