iscsi-target: Remove dead struct iscsi_np_ex logic
This patch removes the unused struct iscsi_np_ex from the core_add_np()
and core_del_np() codepaths, which was originally used for adding an
external network portal entry for SendTargets discovery.

This was dropped during the modern configfs control plane port, and
this patch removes the remaining pieces of unused code.

Signed-off-by: Nicholas A. Bellinger <nab@linux-iscsi.org>