iscsi-target: Convert iscsi_target.c to use 'iscsit_' function prefixes
This patch converts the vast majorty of iscsi_* prefixed functions in
iscsi_target.c code to use 'iscsit_' function prefixes.  At this point
the only functions that are left in iscsi_target.c are the module_init()
+module_exit() calls and iscsi_target_[r,t]x_thread().

This also includes changes this in external code and removes a handful
of dead function prototypes from external includes.

Reported-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Nicholas A. Bellinger <nab@linux-iscsi.org>