target: Make t_task a member of se_cmd, not a pointer
t_task_backstore is a member of se_cmd, and t_task is initialized to
points to it in init_se_cmd. Thus, t_task will never be NULL. Convert
t_task to the member, and eliminate t_task_backstore.

This lets us remove all checks for if se_cmd->se_task is NULL, as well.

Signed-off-by: Andy Grover <agrover@redhat.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>