target: Embed qr in struct se_cmd
This saves a call to kmalloc, and means we don't need to remember to
free the struct se_queue_req.

This also makes it possible to make transport_add_cmd_to_queue never
return an error code, which is nice since it wasn't being checked.

Removed an incorrect comment above get_qr_from_queue; cmd_queue_lock
is acquired in the function now, not before.

Signed-off-by: Andy Grover <agrover@redhat.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>