target: inline struct se_transport_task into struct se_cmd
Continuing the previous patch to make se_transport_task a member
of se_cmd, this removes the nested se_transport_task struct.
There doesn't appear to be any need for a second struct, and
combining the two may let us find redundancies between the two
that we can do without.

Signed-off-by: Andy Grover <agrover@redhat.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>