iscsi-target: Fix incorrect list_for_each_entry() NULL checks
This patch fixes a number of locations where pre struct list_head logic
was incorrectly assuming that a NULL check for *pos after walking the
struct list_head with list_for_each_entry() would indicate that no
matching element was found.

In most of the cases this patch simply converts the matching conditional
break to unlock and return for the matching *pos, and for one special case
in iscsit_handle_status_snack() uses a local int to signal when no *cmd
can be located.

Reported-by: Arne Redlich <arne.redlich@googlemail.com>
Signed-off-by: Nicholas A. Bellinger <nab@linux-iscsi.org>