initramfs: disable event_idx for virtio-scsi-pci
The new event_idx virtio feature does not work for virtio-scsi.  The
first virtqueue kick happens but afterwards the guest thinks it does not
need to kick the host anymore.

This could be an issue in the virtio_scsi guest driver, QEMU virtio-scsi
device emulation, or tcm_vhost.  Simply disable event_idx for now and
debug this later.

Signed-off-by: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>