target/iscsi: Simplify two small bits
I came across these in the paren purge, but wanted to break them out in
a separate patchset so they didn't get lost in the sea of boringness that
those csets embody.

The first bit decrements np->np_exports outside the conditional, and
removes the double-negativity, for clarity.

The second bit removes the test for &cmd->se_cmd entirely, since if cmd
is valid, there's no way the test will ever fail.

Signed-off-by: Andy Grover <agrover@redhat.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>