target: Make transport_generic_get_mem() explictly memset allocated pages
This patch makes transport_generic_get_mem() explictly memset allocated pages in
order to address some control CDB payload strangeness during recent testing.
Be paranoid here and always zero the complete payload just to be sure.

This reverts part of the following commit that removed the memset of
data payloads:

commit 46c20ab9d7c94fd6cf35d41eab7b480127fb0ccf
Author: Andy Grover <agrover@redhat.com>
Date:   Tue May 3 14:48:10 2011 -0700

    target: Don't pass dma_size to generic_get_mem. Don't memset.

This also fixes the failure path that was also dropped in commit:

commit 84b6dd77a6035dff105f628589b8aac8a227031e
Author: Andy Grover <agrover@redhat.com>
Date:   Fri Jul 8 15:22:14 2011 -0700

    target: Eliminate usage of struct se_mem

Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>