tcm_fc: Handle DDP/SW fc_frame_payload_get failures in ft_recv_write_data
Fix ft_recv_write_data() to properly handle fc_frame_payload_get to
return pointer to payload if it exist. If there is no payload which is
most common case (+ve case in case if DDP is working as expected),
it will return NULL. Yes, scope of buf is limited to printk.
Invalidation of HW context (which is done inside ft_invl_hw_context) is
necessary in SUCCESS and FAILURE case of DDP. Hence invalidation is DONE
as long as there was DDP setup (whether it worked correctly or not,
NOTE: For some reason, if there is any error w.r.t DDP ( such as out of
order packet reception), HW simply post the full packet in rx queue))

Problem with moving the 'buf' check will create the problem
(basically doing so will cause DDP invalidation to be not performed in
good case, (FYI: In good case only the last packets header is posted to
rx queue to indicate the completion to the SW so that respective command
can be completed) where as it is necessary to  invalidate DDP context in
HW in good case as well. That 'buf' will be  valid only in case of DDP
failure(error).

Signed-off-by: Kiran Patil <kiran.patil@intel.com>
Cc: Robert W Love <robert.w.love@intel.com>
Signed-off-by: Nicholas A. Bellinger <nab@linux-iscsi.org>