tcm_qla2xxx: Fix confusion about DMA mapping direction
LIO core uses data_direction in the opposite way that qla2xxx needs; for
example, to the LIO core, a READ command is getting data from the target
so it uses DMA_FROM_DEVICE; however for the HBA to be able to DMA the
data to respond to the READ command, the buffer should be mapped with
DMA_TO_DEVICE (since data is flowing from memory to the HBA).

This causes problems on systems with a real IOMMU; eg with VT-d
(intel_iommu) enabled, one sees messages like

    DMAR:[DMA Read] Request device [06:00.0] fault addr ffe4e000
    DMAR:[fault reason 06] PTE Read access is not set

and the target is not able to transfer any data.

Fix this by adding a function tcm_qla2xxx_mapping_dir() that converts
from the LIO core direction to the value that qla2xxx needs to use.

Signed-off-by: Roland Dreier <roland@purestorage.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>