tcm_qla2xxx: Set residual count correctly when we are not sending data
When we return sense data instead of a response, the FC protocol
requires us to set the FCP_RESID_UNDER bit and set FCP_RESID
appropriately (it should be the original FCP_DL value, since we are not
returning any data).  Make sure we set the se_cmd fields appropriately
so qla2xxx does this.

Otherwise, initiators get confused when we fail a command and return a
check condition (eg because of an unimplemented mode page), because we
would leave FCP_RESID as 0 but on the other hand don't transfer any data
in an exchange; this would lead to FC link layer problems.  Sometimes
initiators wouldn't see the sense data at all and would go into a loop
retrying the failing command over and over for a long time.

Signed-off-by: Roland Dreier <roland@purestorage.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>