target: remove the transport_qf_callback se_cmd callback
Remove the need for the transport_qf_callback callback by making
sure we have specific states with specific handlers for the two
queue full cases.

This includes changing a check in the qla2xxx driver to include
TRANSPORT_COMPLETE_QF_WP in addition to TRANSPORT_WRITE_PENDING.  This is
fairly ugly, but not really worse than before.  Long term this leakage
of internal state into drivers needs to go away.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>