qla2xxx: Added a generic IOCB struct typedef in qla_target.h.
This will be used in future patches, for further cleanup of target
mode code.

Since IOCBs are always 64 bytes in size, this struct type can be
used for all ISP types (2xxx, 24xx, etc.), so that some simplifications
can be made associated with the IS_FWI2_CAPABLE() and other related
code.

Signed-off-by: Madhuranath Iyengar <mni@linux-iscsi.org>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>