qla2xxx: Fix a possible hang on target shutdown
If tgt->tgt_stop is set, current code in qla_tgt_do_work() jumps to the
exit path before setting sess, so the code

        if (sess)
                __qla_tgt_sess_put(sess);

skips the kref_put() and therefore sess is be left with an extra
reference forever.  When this happens, qla_tgt_stop_phase1() waits
forever for all the sessions to go away.  Fix this by making sure sess
is always set in qla_tgt_do_work().

Signed-off-by: Roland Dreier <roland@purestorage.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>