qla2xxx: Set cmd->sess / ->loop_id when creating session
Commit 401b01afba24 ("qla2xxx: Remove duplicate assignments from
qla_tgt_do_work") removed setting cmd->sess and cmd->loop_id, which is
usually OK, except when we end up creating a new session in
qla_tgt_do_work(), when it leads to dropping the command with the
message from tcm_qla2xxx_handle_cmd():

    Unable to locate struct qla_tgt_sess from qla_tgt_cmd

Re-add the needed assignments.

Longer term it seems we may want to move the looking up of sessions
into qla_tgt_do_work(), where it will be easier to avoid raciness that
leads to creating multiple local sessions when a burst of commands
comes in.

Signed-off-by: Roland Dreier <roland@purestorage.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>