qla2xxx: Fix sign of FCP_RESID field for status with RESID_OVER
We don't need to change the sign of our prm->residual field; the target
core sets a flag to distinguish overflow and underflow, but
se_cmd->residual_count is always positive.  The existing code seems to
be a carryover from the port from scst, which uses the sign of the
residual field to distinguish overflow and underflow.

Signed-off-by: Roland Dreier <roland@purestorage.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>