staging: line6: drop midi_mask_transmit
The midi_mask_transmit sysfs attribute selects the MIDI channels on
which to transmit messages.  If ALSA wants to transmit a message that
does not match an allowed channel, the message will be dropped.  This
driver feature is not really used and applications don't know how to
take advantage of it.  Therefore we drop it and rely on applications or
MIDI sequencers to select the channels used for communication.

Signed-off-by: Stefan Hajnoczi <stefanha@gmail.com>