eventfd: add ioctl(EVENTFD_SET_POLL_INFO)
Add busy_loop polling support to eventfd.  A userspace memory location
can be registered with the eventfd for busy_loop polling.  The new ioctl
EVENTFD_SET_POLL_INFO is used to set the userspace memory location
details that will be polled.

A typical use case is a descriptor ring index where the eventfd is the
doorbell/notification used to wake the reader.  By polling the
descriptor ring index instead of requiring the eventfd to be signalled
we can reduce latency.

This patch is experimental and incomplete.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>